% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_ht.R
\name{param_ht}
\alias{param_ht}
\title{Generating parameters of the total tree height model from Auger (2016)}
\usage{
param_ht(
  fic_arbres,
  nb_iter = 1,
  mode_simul = "DET",
  nb_step = 1,
  dt = 10,
  seed_value = NULL
)
}
\arguments{
\item{fic_arbres}{Dataframe containing a list of trees for which height must be estimated, at least tree ID (no_arbre) and plot Id (id_pe) must be present}

\item{nb_iter}{Number of iteration if stochastic mode (default 1), ignored if \code{mode_simul="DET"}}

\item{mode_simul}{Simulation mode (STO = stochastic, DET = deterministic), default "DET"}

\item{nb_step}{Number of time steps for which we need height estimation for the same tree (default 1), ignored if \code{mode_simul="DET"}}

\item{dt}{Lenght of time step if nb_step>1 (default 10), ignored if \code{mode_simul="DET"}}

\item{seed_value}{Optionnal parameter to set the seed value for random number generation. Usually used for testing for reproductibility}
}
\value{
A list of lists, one list per species for which there is a tree height model, and for each species, a list with 4 elements:
\enumerate{
  \item essence: string with species code name
  \item effet_fixe: dataframe of fixed effect parameter values (one line per iteration)
  \item random_placette: dataframe of random plot effect values (one line per plot per iteration)
  \item erreur_residuelle: dataframe of residual error values (one line per tree per iteration per time step)
}
}
\description{
Generate parameters of the tree height model from Auger (2016), deterministic or stochastic, for each tree of each plot and for all iterations
}
\details{
The model for estimating total tree height is a linear mixed model calibrated per species (Auger 2016).
There is a random plot effect, correlated residual errors with CorCar1 correlation fonction, and a covariance matrix of fixed effects.

Auger, I., 2016. Une nouvelle relation hauteur-diamètre tenant compte de l’influence de la station et
du climat pour 27 essences commerciales du Québec. Gouvernement du Québec, ministère
des Forêts, de la Faune et des Parcs, Direction de la recherche forestière. Note de recherche forestière no 146. 31 p.
}
\examples{
# Mode déterministe
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test)

# Mode stochastique, pour un seul pas de simulation et 10 itérations
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10)

# Mode stochastique, plusieurs pas de simulation et plusieurs itérations
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10, nb_step=5)

}
