% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_vol.R
\name{param_vol}
\alias{param_vol}
\title{Génère les paramètres pour estimer le volume marchand sans écorce d'un arbre avec l'équation de Fortin et al. (2007)}
\usage{
param_vol(
  fic_arbres,
  mode_simul = "DET",
  nb_iter = 1,
  nb_step = 1,
  seed_value = NULL
)
}
\arguments{
\item{fic_arbres}{Une table contenant la liste d'arbres dont le volume est à estimer. La table doit contenir les variables \code{no_arbre} (l'identifiant de l'arbre) et \code{id_pe} (l'identifiant de la placette).}

\item{mode_simul}{Le mode de simulation (STO = stochastiqie, DET = déterministe), par défaut "DET".}

\item{nb_iter}{Le nombre d'itérations si le mode stochastique est utilisé, doit être > 1. Ignoré si \code{mode_simul="DET"}.}

\item{nb_step}{Le nombre d'années pour lesquelles on veut estimer le volume pour un même arbre (par défaut 1), ignoré si \code{mode_simul="DET"}.}

\item{seed_value}{La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction. Optionnel.}
}
\value{
La fonction retourne une liste de listes, soit une liste par itération. Pour une itération, la liste contient 3 éléments:

\enumerate{
  \item effet_fixe : une table avec 5 colonnes: 3 pour les paramètres des effets fixes de l'équation, numéro de l'itération et essence. Une ligne par essence.
  \item random_placette: une table avec 3 colonnes: effet aléatoire de placette, le numéro de l'itération et l'identifiant de la placette. Une ligne par placette/itération/step. 0 si \code{mode_simul="DET"}.
  \item erreur_residuelle : une table avec l'erreur résiduelle associée à chaque essence (une colonne par essence), identifiant de la placette, identifiant de l'arbre, et numéro de l'itération. Une ligne par arbre/placette/itération. 0 si \code{mode_simul="DET"}.
}
}
\description{
Génère les paramètres pour estimer le volume marchand sans écorce d'un arbre avec l'équation de Fortin et al. (2007)
Les paramètres peuvent être générés de façon déterministe ou stochastique.
Si le mode stochastique est utilisé, les paramètres seront générés pour tous les arbres/itétations/années.
}
\details{
L'équation pour estimer le volume marchand a été étalonnée avec un modèle linéaire mixte où l'essence est une covariable dans l'équation (Fortin et al. 2007).
Le modèle inclut un effet aléatoire de placette et un effet aléatoire de virée, mais seulement l'effet aléatoire de placette est simulé.

Fortin, M., J. DeBlois, S. Bernier et G. Blais, 2007. Mise au point d’un tarif de cubage général pour
les forêts québécoises : une approche pour mieux évaluer l’incertitude associée aux prévisions.
For. Chron. 83: 754-765.
}
\examples{
# Mode déterministe
parametre_vol <- param_vol(fic_arbres=fic_arbres_test)

# Mode stochastique, pour une seule année et 10 itérations
parametre_vol <- param_vol(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10)

#' # Mode stochastique, plusieurs années et 10 itérations
parametre_vol <- param_vol(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10, nb_step=3)

}
