% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifQC-package.R
\docType{package}
\name{TarifQC-package}
\alias{TarifQC}
\alias{TarifQC-package}
\title{TarifQC: Estimating total tree height and merchantable volume}
\description{
Estimating total tree height with the model from Auger (2016) and estimating tree merchantable volume with the model from Fortin et al. (2007). Estimation can be stochastic or deterministic.
}
\keyword{internal}
