% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_ht.R
\name{param_ht}
\alias{param_ht}
\title{Génère les paramètres pour estimer la hauteur totale d'un arbre avec l'équation de Auger (2016)}
\usage{
param_ht(
  fic_arbres,
  mode_simul = "DET",
  nb_iter = 1,
  nb_step = 1,
  dt = 10,
  seed_value = NULL
)
}
\arguments{
\item{fic_arbres}{Une table contenant la liste d'arbres dont la hauteur est à estimer. La table doit contenir les variables \code{no_arbre} (l'identifiant de l'arbre) et \code{id_pe} (l'identifiant de la placette).}

\item{mode_simul}{Le mode de simulation (STO = stochastique, DET = déterministe), par défaut "DET".}

\item{nb_iter}{Le nombre d'itérations si le mode stochastique est utilisé, doit être > 1. Ignoré si \code{mode_simul="DET"},}

\item{nb_step}{Le nombre d'années pour lesquelles on veut estimer la hauteur pour un même arbre (par défaut 1), ignoré si \code{mode_simul="DET"}.}

\item{dt}{La durée de l'intervalle de temps entre deux mesures d'un même arbre si \code{nb_step>1} (par défaut 10), ignoré si \code{mode_simul="DET"}.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}
}
\value{
La fonction retourne une liste de listes, soit une liste par essence pour laquelle il y a une équation de hauteur.
Pour une essence, la liste contient 4 éléments:
\enumerate{
  \item essence: le code de l'essence
  \item effet_fixe: une table avec une colonne pour chacun des paramètres des effets fixes de l'équation et une ligne par iteration. Une seule ligne si \code{mode_simul="DET"}.
  \item random_placette: une table avec 4 colonnes: effet aléatoire de placette, le numéro de l'itération, l'identifiant de la placette et le code de l'essence. Une ligne par placette/itération. 0 si \code{mode_simul="DET"}.
  \item erreur_residuelle: une table avec 6 colonnes:  erreur résiduelle, identifiant de la placette, identifiant de l'arbre, numéro de step, numéro de l'itération et le code de l'essence. Une ligne par arbre/placette/itération/step. 0 si \code{mode_simul="DET"}.
}
}
\description{
Génère les paramètres pour estimer la hauteur totale d'un arbre avec l'équation de Auger (2016).
Les paramètres peuvent être générés de façon déterministe ou stochastique.
Si le mode stochastique est utilisé, les paramètres seront générés pour tous les arbres/itétations/années.
}
\details{
L'équation pour estimer la hauteur totale a été étalonnée avec un modèle linéaire mixte, par essence (Auger 2016).
Le modèle inclut un effet aléatoire de placette et une corrélation de type CorCar1 sur les erreurs résiduelles.

Auger, I., 2016. Une nouvelle relation hauteur-diamètre tenant compte de l’influence de la station et
du climat pour 27 essences commerciales du Québec. Gouvernement du Québec, ministère
des Forêts, de la Faune et des Parcs, Direction de la recherche forestière. Note de recherche forestière no 146. 31 p.
}
\examples{
# Mode déterministe
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test)

# Mode stochastique, pour une seule année et 10 itérations
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10)

# Mode stochastique, plusieurs années et plusieurs itérations
parametre_ht_dhp <- param_ht(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10, nb_step=5)

}
