% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_vol.R
\name{param_vol}
\alias{param_vol}
\title{Generating parameters of the merchantable tree volume model from Fortin et al. (2007)}
\usage{
param_vol(fic_arbres, mode_simul = "DET", nb_iter = 1, seed_value = NULL)
}
\arguments{
\item{fic_arbres}{Dataframe with a list of trees for which volume must be estimated, at least tree ID (no_arbre) and plot Id (id_pe) must be present}

\item{mode_simul}{Simulation mode (STO = stochastic, DET = deterministic), default "DET"}

\item{nb_iter}{Number of iterations if stochastic mode (default 1), ignored if \code{mode_simul="DET"}}

\item{seed_value}{Optionnal parameter to set the seed value for random number generation. Usually used for testing for reproductibility}
}
\value{
A list of lists, one list per iteration, and for each iteration, a list with 3 elements:
\enumerate{
  \item effet_fixe : dataframe of fixed effect parameter values (one line per species for which there is a volume model)
  \item random_placette: dataframe of random plot effect values (one line per plot)
  \item erreur_residuelle : dataframe of residual error values (one line per tree)
}
}
\description{
Generate parameters of the merchantable tree volume model of Fortin et al. (2007), deterministic or stochastic, for each tree of each plot and for all iterations
}
\details{
The model for estimating merchantable tree volume is a linear mixed model with species as a covariate (Fortin et al. 2007).
There is a random plot effect and a viree random effect, but only the plot random effect is generated.
The residual errors are not correlated. There is a covariance matrix of fixed effects.

Fortin, M., J. DeBlois, S. Bernier et G. Blais, 2007. Mise au point d’un tarif de cubage général pour
les forêts québécoises : une approche pour mieux évaluer l’incertitude associée aux prévisions.
For. Chron. 83: 754-765.
}
\examples{
# Mode déterministe
parametre_vol <- param_vol(fic_arbres=fic_arbres_test)

# Mode stochastique, 10 itérations
parametre_vol <- param_vol(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10)

}
