% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relation_hauteur_diametre.R
\name{relation_h_d}
\alias{relation_h_d}
\title{Estime la hauteur totale de chacun des arbres avec l'équation de Auger (2016).}
\usage{
relation_h_d(
  fic_arbres,
  mode_simul = "DET",
  nb_iter = 1,
  nb_step = 1,
  dt = 10,
  seed_value = NULL,
  grouping_vars = NULL,
  reg_eco = FALSE
)
}
\arguments{
\item{fic_arbres}{Une table contenant la liste d'arbres regroupés au minimum en placettes avec les informations suivantes:
\itemize{
   \item id_pe: identifiant unique de la placette
   \item dhpcm: dhp (cm) de l'arbre ou classe de dhp (>9 cm)
   \item essence: code d'essence de l'arbre (ex: SAB, EPN, BOP)
   \item no_arbre: identifiant de l'arbre ou de la combinaison dhp/essence, nécessaire seulement si \code{mode_simul}='STO'
   \item nb_tige: nombre d'arbres de l'essence et de la classe de dhp, dans 400 m2 (pour calculer la surface terrière de la placette)
   \item sdom_bio: sous-domaine bioclimatique, en majuscule (ex: 1, 2E, 4O), seuls les domaines 1 à 6 sont traités
   \item veg_pot: code de végétation potentielle, 3 premiers caractères du type écologique  (ex: FE3, MS2)
   \item milieu: 4e caractère du type écologique, doit être caractère, une valeur de 0 à 9 (ex: 2)
   \item p_tot: précipitation totale annuelle moyenne sur la période 1980-2010 (mm)
   \item t_ma: température annuelle moyenne sur la période 1980-2010 (Celcius)
   \item altitude: altitude (m)
   \item reg_eco: Optionnel. Code de la région écologique. Vous pouvez fournir la région au lieu du sous-domaine, alors mettre le paramètre \code{reg_eco=TRUE}
   \item iter: numéro de l'itération, seulement si mode stochastique, doit être numéroté de 1 à nb_iter
   \item step: numéro de la step, seulement si mode stochastique, doit être numéroté de 1 à nb_step. Obligatoire même si le fichier n'est qu'une liste d'arbres à un moment donné.
}}

\item{mode_simul}{Mode de simulation : STO = stochastistique, DET = déterministe), par défaut "DET"}

\item{nb_iter}{Le nombre d'itérations si le mode stochastique est utilisé, doit être > 1. Ignoré si \code{mode_simul="DET"},}

\item{nb_step}{Le nombre d'années pour lesquelles on veut estimer la hauteur pour un même arbre (par défaut 1), ignoré si \code{mode_simul="DET"}.}

\item{dt}{La durée de l'intervalle de temps entre deux mesures d'un même arbre si \code{nb_step>1} (par défaut 10), ignoré si \code{mode_simul="DET"}.}

\item{seed_value}{Optionnel. La valeur du seed pour la génération de nombres aléatoires. Généralement utilisé pour les tests de la fonction.}

\item{grouping_vars}{Optionel. Si \code{mode_simul}='DET', les colonnes à ajouter comme variables de groupement, en plus de id_pe, pour calculer la surface terrière d'une placette.
Par exemple, si le fichier des arbres contient plus d'une année par arbre, ajouter le colonne identifiant l'année comme variable de groupement: grouping_vars='var1'.
S'il y a plusieurs variables de groupement: grouping_vars=c('var1','var2').}

\item{reg_eco}{Optionel. Mettre à \code{TRUE} si reg_eco est fourni dans \code{fic_arbres} au lieu de sdom_bio. reg_eco sera converti en sdom_bio. La colonne sdom_bio ne doit pas être dans \code{fic_arbres}.}
}
\value{
La table \code{fic_arbres} avec une colonne contenant la hauteur estimée en mètres (hauteur_pred).
}
\description{
Estime la hauteur totale en mètre de chacun des arbres avec l'équation de Auger (2016). La fonction permet l'estimation pour une liste d'arbres regroupés en placette.
L'estimation peut être déterministe ou stochastique.
}
\details{
L'équation pour estimer la hauteur totale a été étalonnée avec un modèle linéaire mixte, par essence (Auger 2016).
Le modèle inclut un effet aléatoire de placette et une corrélation de type CorCar1 sur les erreurs résiduelles.
La fonction estime la hauteur de façon déterministe ou stochastique. Si \code{mode_simul}='STO', les paramètres doivent être générés préalablement avec la fonction \code{param_ht}.
Voir les exemples.

Auger, I., 2016. Une nouvelle relation hauteur-diamètre tenant compte de l’influence de la station et
du climat pour 27 essences commerciales du Québec. Gouvernement du Québec, ministère
des Forêts, de la Faune et des Parcs, Direction de la recherche forestière. Note de recherche forestière no 146. 31 p.
}
\examples{
# Exemple 1: DETERMINISTE: un seule année par arbre ---------------------------------------
DataHt <- relation_h_d(fic_arbres=fic_arbres_test)

# Exemple 2: DETERMINISTE: avec grouping_vars, plusieurs années par arbre ----------------
DataHt <- relation_h_d(fic_arbres=fic_artemis_det, grouping_vars='annee')

# Exemple 3: STOCHASTIQUE: plusieurs années par arbre ------------------------------------
nb_iter <- length(unique(fic_artemis_sto$iter))
nb_step <- length(unique(fic_artemis_sto$step))
DataHt <- relation_h_d(fic_arbres=fic_artemis_sto, mode_simul='STO', nb_iter=nb_iter, nb_step=nb_step)

}
