% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relation_hauteur_diametre.R
\name{relation_h_d}
\alias{relation_h_d}
\title{Estimating total tree height from DBH and stand characteristics for a list of trees in plots}
\usage{
relation_h_d(
  fic_arbres,
  mode_simul = "DET",
  iteration = 1,
  step = 1,
  parametre_ht = NULL,
  grouping_vars = NULL
)
}
\arguments{
\item{fic_arbres}{Dataframe containing a list trees in plots, with DBH and stand characteristics:
\itemize{
   \item id_pe: identifiant unique de la placette
   \item dhpcm: dhp (cm) de l'arbre ou classe de dhp (>9 cm)
   \item essence: code d'essence de l'arbre (ex: SAB, EPN, BOP)
   \item no_arbre: identifiant de l'arbre ou de la combinaison dhp/essence, nécessaire seulement si \code{mode_simul}='STO'
   \item nb_tige: nombre d'arbres de l'essence et de la classe de dhp, dans 400 m2
   \item sdom_bio: sous-domaine bioclimatique, en majuscule (ex: 2E, 4O), domaines 1 à 6 sont traités
   \item veg_pot: code de végétation potentielle, 3 premiers caractères du type écologique  (ex: FE3, MS2)
   \item milieu: 4e caractère du type écologique, doit être numérique, une valeur de 0 à 9 (ex: 2)
   \item p_tot: 1980-2010 mean total precipitations (mm)
   \item t_ma: 1980-2010 mean annual temperature (Celcius)
   \item altitude: altitude en m
}}

\item{mode_simul}{Simulation mode (STO = stochastic, DET = deterministic), default "DET"}

\item{iteration}{If \code{mode_simul}='STO', iteration number to estime (default 1)}

\item{step}{If \code{mode_simul}='STO', Time step to estimate (default 1)}

\item{parametre_ht}{If \code{mode_simul}='STO', oject with model parameters values pour all iterations and time steps, provided by \code{param_ht} function. If \code{mode_simul}='DET', parameter values are prepared directly in this function, but \code{param_ht} can also be used.}

\item{grouping_vars}{If \code{mode_simul}='DET', columns to be added as grouping variables in addition to id_pe variable for calculating basal area of the plots. For example, if there is multiple measurements, add measurement id as grouping variable: grouping_vars='var1'. If multiple grouping variables, use grouping_vars=c('var1','var2'). Optional. If \code{mode_simul}='STO', this parameter can not be used. In this case, concatenate all the grouping variables in one variable id_pe}
}
\value{
Dataframe \code{fic_arbres} with column hauteur_pred with estimated height in meter
}
\description{
Estimate total tree height in meter with Auger (2016) equation for a list of trees in plots in a data frame.
Estimation can be deterministic or stochastic.
}
\details{
The models for estimating tree height are linear mixed models calibrated by species (Auger 2016).
There is a random plot effect and correlated residual errors with CorCar1 correlation fonction.
This function estimates height, using deterministic parameters or stochastic parameters, for a given iteration and time step.
If \code{mode_simul}='STO', the parameters must be loaded with \code{param_ht} function before using this function.

Auger, I., 2016. Une nouvelle relation hauteur-diamètre tenant compte de l’influence de la station et
du climat pour 27 essences commerciales du Québec. Gouvernement du Québec, ministère
des Forêts, de la Faune et des Parcs, Direction de la recherche forestière. Note de recherche forestière no 146. 31 p.
}
\examples{
# Exemple 1: DETERMINISTE: un seul mesurage par arbre ---------------------------------------
DataHt <- relation_h_d(fic_arbres=fic_arbres_test)

# Exemple 2: DETERMINISTE: avec grouping_vars, plusieurs mesurages par arbre ----------------
DataHt <- relation_h_d(fic_arbres=fic_artemis_det, grouping_vars='step')

# Exemple 3: STOCHASTIQUE: un seul mesurage par arbre ---------------------------------------
# Générer les paramètres pour plusieurs itérations
parametre_ht <- param_ht(fic_arbres=fic_arbres_test, mode_simul='STO', nb_iter=10)
# Estimer la hauteur pour l'itération 2
DataHt <- relation_h_d(fic_arbres=fic_arbres_test, mode_simul='STO', iteration=2, parametre_ht=parametre_ht)

# Exemple 4: STOCHASTIQUE: plusieurs mesurages par arbre ------------------------------------
# Générer les paramètres pour plusieurs itérations et steps
parametre_ht <- param_ht(fic_arbres=fic_artemis_sto, mode_simul='STO', nb_iter=10, nb_step=5)
# Estimer la hauteur pour l'itération 2 et la step 3
DataHt <- relation_h_d(fic_arbres=fic_artemis_sto[fic_artemis$iter==10 & fic_artemis$step==3,], mode_simul='STO', iteration=2, step=3, parametre_ht=parametre_ht)

# Exemple 5: STOCHASTIQUE: traiter toutes les itérations et steps ---------------------------
# Générer les paramètres du modèle de hauteur pour toutes les itérations et time steps
nb_iter <- length(unique(fic_artemis_sto$iter)) # 10
nb_step <- length(unique(fic_artemis_sto$step)) # 5
parametre_ht <- param_ht(fic_arbres=fic_artemis_sto, mode_simul='STO', nb_iter=nb_iter, nb_step=nb_step)
# Appliquer le modèle de hauteur à chaque iteration/step
fic_artemis_final1 <- NULL
for (i in 1:nb_iter){
  for (k in 1:nb_step){
      ht <- relation_h_d(fic_arbres=fic_artemis_sto[fic_artemis_sto$iter==i & fic_artemis_sto$step==k,], mode_simul='STO', iteration=i, step=k, parametre_ht=parametre_ht)
      fic_artemis_final1 <- bind_rows(fic_artemis_final1, ht)
      }
   }

# On peut aussi paralléliser les deux boucles for
nb_iter <- length(unique(fic_artemis_sto$iter)) # 10
nb_step <- length(unique(fic_artemis_sto$step)) # 5
parametre_ht <- param_ht(fic_arbres=fic_artemis_sto, mode_simul='STO', nb_iter=nb_iter, nb_step=nb_step)
# Appliquer le modèle de hauteur à chaque iteration/step
registerDoFuture()
plan(multisession)
fic_artemis_final2 <- bind_rows(
  foreach (i = 1:nb_iter) \%:\% # nesting operator
      foreach (k = 1:nb_step) \%dopar\% {
            fic <- relation_h_d(fic_arbres=fic_artemis_sto[fic_artemis$iter==i & fic_artemis$step==k,], mode_simul='STO', iteration=i, step=k, parametre_ht=parametre_ht)
            }
 )
}
